# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: MBN AI 트롯 팬덤 플랫폼 API
  version: 0.1.0
  description: |
    FE/BE 공통 계약. **이 파일이 단일 진실 공급원입니다.**

    - BE: 이 스펙대로 컨트롤러를 구현합니다. 스펙에 없는 엔드포인트를 임의로 추가하지 마세요.
    - FE: `npm run api:types` 로 이 파일에서 TS 타입을 생성합니다. 타입을 손으로 쓰지 마세요.
    - 계약 변경은 **이 파일 수정 → 양쪽 반영** 순서로만 진행합니다.

    도메인 정의는 `docs/domain-model.md` 참조.
servers:
  - url: http://localhost:8080
    description: 로컬 개발
  - url: /
    description: 동일 오리진 배포
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/schedules:
    get:
      tags: [Schedule]
      summary: 일정 목록
      security: []
      parameters:
        - $ref: '#/components/parameters/StarIdQuery'
        - name: upcoming
          in: query
          description: true면 현재 시각 이후 일정만
          schema: { type: boolean, default: true }
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/ScheduleType' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/Schedule' }
  /api/v1/schedules/{id}:
    get:
      tags: [Schedule]
      summary: 일정 상세
      security: []
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Schedule' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────────── Content ──────────────────────────
...
  /api/v1/gatherings:
    get:
      tags: [Gathering]
      summary: 팬 모임 목록
      security: []
      parameters:
        - $ref: '#/components/parameters/StarIdQuery'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/GatheringStatus' }
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/GatheringType' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/GatheringSummary' }
  /api/v1/gatherings/{id}:
    get:
      tags: [Gathering]
      summary: 팬 모임 상세
      description: 인증된 사용자의 경우 `myApplication` 필드로 본인 신청 상태를 함께 반환합니다.
      security:
        - {}
        - bearerAuth: []
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Gathering' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/gatherings/{id}/applications:
    post:
      tags: [Gathering]
      summary: 모임 참여 신청
      description: |
        `currentCount`를 원자적으로 증가시킵니다. 동시 신청 경합을 반드시 처리해야 합니다.
        중복 신청 시 409, 정원 초과 시 422를 반환합니다.
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string, maxLength: 200 }
      responses:
        '201':
          description: 신청 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GatheringApplication' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 이미 신청한 모임
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: 정원 초과 또는 모집 마감
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/gatherings/{id}/applications/me:
    delete:
      tags: [Gathering]
      summary: 모임 참여 신청 취소
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '204': { description: 취소 완료 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ────────────────────────── Place ───────────────────────────
  /api/v1/places:
    get:
      tags: [Place]
      summary: 스타가 다녀간 장소 목록
      security: []
      parameters:
        - $ref: '#/components/parameters/StarIdQuery'
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/PlaceType' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/Place' }
  /api/v1/places/{id}:
    get:
      tags: [Place]
      summary: 장소 상세
      security: []
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Place' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─────────────────────────── Tip ────────────────────────────
  /api/v1/tips:
    get:
      tags: [Tip]
      summary: 팬덤 팁 목록
      security: []
      parameters:
        - $ref: '#/components/parameters/StarIdQuery'
        - name: category
          in: query
          schema: { $ref: '#/components/schemas/TipCategory' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/TipSummary' }
  /api/v1/tips/{id}:
    get:
      tags: [Tip]
      summary: 팬덤 팁 상세
      security: []
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tip' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ────────────────────────── Search ──────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    StarIdQuery:
      name: starId
      in: query
      required: true
      schema: { type: integer, format: int64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: GATHERING_ALREADY_APPLIED }
        message: { type: string, example: 이미 신청한 모임입니다. }
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages, last]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        last: { type: boolean }
...
    ScheduleType:
      type: string
      enum: [BROADCAST, CONCERT, FANMEETING, EVENT]
    Schedule:
      type: object
      required: [id, title, type, startAt, official]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string, example: 임영웅의 팬미팅 }
        type: { $ref: '#/components/schemas/ScheduleType' }
        startAt: { type: string, format: date-time, example: '2026-08-07T10:00:00Z' }
        endAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        venue: { type: string }
        description: { type: string }
        official: { type: boolean }
        externalUrl:
          oneOf:
            - { type: string, format: uri }
            - { type: 'null' }

    # ── Content (기사 · 영상) ──
...
    GatheringStatus:
      type: string
      enum: [RECRUITING, FULL, CLOSED]
    GatheringType:
      type: string
      enum: [BUS, DONATION, AD, GROUP_VIEWING, ETC]
    GatheringSummary:
      type: object
      required: [id, title, status, coverImageUrl, currentCount, capacity, deadline, hostNickname, official]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string, example: 0729 관광 버스 대절 모집 (대전 출발) }
        type: { $ref: '#/components/schemas/GatheringType' }
        status: { $ref: '#/components/schemas/GatheringStatus' }
        coverImageUrl: { type: string, format: uri }
        summary:
          type: string
          description: 카드에 2줄로 노출되는 요약
          example: 대전역에서 콘서트장까지 가는 버스입니다. 08:00까지 도착하시면 됩니다.
        currentCount: { type: integer, example: 31 }
        capacity: { type: integer, example: 40 }
        deadline: { type: string, format: date, example: '2026-08-03' }
        hostNickname: { type: string, example: 임영웅 사생팬 }
        official: { type: boolean, description: MBN 검수 배지 노출 여부 }
    Gathering:
      allOf:
        - $ref: '#/components/schemas/GatheringSummary'
        - type: object
          required: [description, eventAt, meetingPoint, fee]
          properties:
            description: { type: string }
            eventAt: { type: string, format: date-time }
            meetingPoint: { type: string, example: 대전역 }
            fee:
              type: integer
              description: 참가 비용(원). 0이면 무료. **표시 전용 — 플랫폼은 결제를 중개하지 않습니다.**
            paymentInfo: { type: string }
            refundPolicy: { type: string }
            notice: { type: string }
            myApplication:
              description: 인증된 사용자의 본인 신청 상태. 비인증 또는 미신청 시 null.
              oneOf:
                - $ref: '#/components/schemas/GatheringApplication'
                - type: 'null'
    GatheringApplication:
      type: object
      required: [id, gatheringId, status, appliedAt]
      properties:
        id: { type: integer, format: int64 }
        gatheringId: { type: integer, format: int64 }
        status: { type: string, enum: [APPLIED, CANCELED] }
        note: { type: string }
        appliedAt: { type: string, format: date-time }

    # ── Place ──
    PlaceType:
      type: string
      enum: [RESTAURANT, CAFE, VENUE, FILMING_LOCATION]
    Place:
      type: object
      required: [id, name, type, address, imageUrl, sourceUrl]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, example: 한정식 서울역점 }
        type: { $ref: '#/components/schemas/PlaceType' }
        address: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        imageUrl: { type: string, format: uri }
        visitContext: { type: string, description: 스타가 방문한 프로그램 또는 시점 }
        sourceUrl:
          type: string
          format: uri
          description: '**필수** — 공개된 출처만 등록. 실시간 위치·사적 동선은 금지.'
        mapUrl:
          oneOf: [{ type: string, format: uri }, { type: 'null' }]
          description: 카카오맵 외부 링크. 지도를 임베드하지 않습니다.
        relatedContentIds:
          type: array
          items: { type: integer, format: int64 }

    # ── Tip ──
    TipCategory:
      type: string
      enum: [VOTE, STREAMING, TICKETING, CHEER, FANCLUB]
    TipSummary:
      type: object
      required: [id, title, category, thumbnailUrl, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string, example: 멜론 투표 방법 }
        category: { $ref: '#/components/schemas/TipCategory' }
        thumbnailUrl: { type: string, format: uri }
        updatedAt:
          type: string
          format: date-time
          description: 플랫폼 정책이 자주 바뀌므로 화면에 **항상 노출**합니다.
    Tip:
      allOf:
        - $ref: '#/components/schemas/TipSummary'
        - type: object
          required: [content]
          properties:
            content: { type: string, description: 마크다운 }
            externalUrl:
              oneOf:
                - { type: string, format: uri }
                - { type: 'null' }

    # ── Chat ──
...